Issue #11163: Split InputHiddenField3 into focused test inputs
