+++ src/test/java/com/puppycrawl/tools/checkstyle/checks/coding/HiddenFieldCheckTest.java
@@ -29,7 +29,6 @@ import java.util.function.Predicate;
 import org.junit.jupiter.api.Test;
 
 import com.puppycrawl.tools.checkstyle.AbstractModuleTestSupport;
-import com.puppycrawl.tools.checkstyle.DefaultConfiguration;
 import com.puppycrawl.tools.checkstyle.JavaParser;
 import com.puppycrawl.tools.checkstyle.api.DetailAST;
 import com.puppycrawl.tools.checkstyle.api.TokenTypes;
@@ -231,48 +230,80 @@ public class HiddenFieldCheckTest
                  getPath("InputHiddenField2Misc.java"), expected);
     }
 
-    /** Tests ignoreFormat property. */
     @Test
-    public void testIgnoreFormat()
-            throws Exception {
-        final DefaultConfiguration checkConfig =
-            createModuleConfig(HiddenFieldCheck.class);
-        checkConfig.addProperty("ignoreFormat", "^i.*$");
-        assertWithMessage("Ignore format should not be null")
-            .that(checkConfig.getProperty("ignoreFormat"))
-            .isNotNull();
+    public void testHiddenField3Basic() throws Exception {
         final String[] expected = {
-            "30:13: " + getCheckMessage(MSG_KEY, "hidden"),
-            "33:34: " + getCheckMessage(MSG_KEY, "hidden"),
-            "40:13: " + getCheckMessage(MSG_KEY, "hidden"),
-            "45:18: " + getCheckMessage(MSG_KEY, "hidden"),
-            "50:33: " + getCheckMessage(MSG_KEY, "hidden"),
-            "71:17: " + getCheckMessage(MSG_KEY, "hidden"),
-            "79:22: " + getCheckMessage(MSG_KEY, "hidden"),
+            "21:13: " + getCheckMessage(MSG_KEY, "hidden"),
+            "24:39: " + getCheckMessage(MSG_KEY, "hidden"),
+            "31:13: " + getCheckMessage(MSG_KEY, "hidden"),
+            "36:18: " + getCheckMessage(MSG_KEY, "hidden"),
+            "41:33: " + getCheckMessage(MSG_KEY, "hidden"),
+            "62:17: " + getCheckMessage(MSG_KEY, "hidden"),
+            "71:22: " + getCheckMessage(MSG_KEY, "hidden"),
+            "78:17: " + getCheckMessage(MSG_KEY, "hidden"),
             "85:17: " + getCheckMessage(MSG_KEY, "hidden"),
-            "92:17: " + getCheckMessage(MSG_KEY, "hidden"),
-            "97:13: " + getCheckMessage(MSG_KEY, "hidden"),
-            "115:29: " + getCheckMessage(MSG_KEY, "prop"),
-            "121:29: " + getCheckMessage(MSG_KEY, "prop"),
-            "127:29: " + getCheckMessage(MSG_KEY, "prop"),
-            "139:28: " + getCheckMessage(MSG_KEY, "prop"),
-            "153:13: " + getCheckMessage(MSG_KEY, "hidden"),
-            "158:13: " + getCheckMessage(MSG_KEY, "hidden"),
-            "163:13: " + getCheckMessage(MSG_KEY, "hidden"),
-            "167:13: " + getCheckMessage(MSG_KEY, "hidden"),
-            "194:23: " + getCheckMessage(MSG_KEY, "y"),
-            "215:17: " + getCheckMessage(MSG_KEY, "hidden"),
-            "225:22: " + getCheckMessage(MSG_KEY, "hidden"),
-            "232:13: " + getCheckMessage(MSG_KEY, "hidden"),
-            "238:13: " + getCheckMessage(MSG_KEY, "hiddenStatic"),
-            "245:41: " + getCheckMessage(MSG_KEY, "x"),
-            "251:30: " + getCheckMessage(MSG_KEY, "xAxis"),
-            "268:41: " + getCheckMessage(MSG_KEY, "prop"),
-            "282:29: " + getCheckMessage(MSG_KEY, "prop"),
-            "293:42: " + getCheckMessage(MSG_KEY, "prop2"),
+            "90:13: " + getCheckMessage(MSG_KEY, "hidden"),
         };
         verifyWithInlineConfigParser(
-                getPath("InputHiddenField3.java"), expected);
+                 getPath("InputHiddenField3Basic.java"), expected);
+    }
+
+    @Test
+    public void testHiddenField3Interface() throws Exception {
+        final String[] expected = {
+        };
+        verifyWithInlineConfigParser(
+                 getPath("InputHiddenField3Interface.java"), expected);
+    }
+
+    @Test
+    public void testHiddenField3Static() throws Exception {
+        final String[] expected = {
+            "22:13: " + getCheckMessage(MSG_KEY, "hidden"),
+            "27:13: " + getCheckMessage(MSG_KEY, "hidden"),
+            "32:13: " + getCheckMessage(MSG_KEY, "hidden"),
+            "36:13: " + getCheckMessage(MSG_KEY, "hidden"),
+            "63:23: " + getCheckMessage(MSG_KEY, "y"),
+        };
+        verifyWithInlineConfigParser(
+                 getPath("InputHiddenField3Static.java"), expected);
+    }
+
+    @Test
+    public void testHiddenField3PropertySetter() throws Exception {
+        final String[] expected = {
+            "21:29: " + getCheckMessage(MSG_KEY, "prop"),
+            "27:29: " + getCheckMessage(MSG_KEY, "prop"),
+            "33:29: " + getCheckMessage(MSG_KEY, "prop"),
+            "45:28: " + getCheckMessage(MSG_KEY, "prop"),
+            "64:41: " + getCheckMessage(MSG_KEY, "prop"),
+            "78:29: " + getCheckMessage(MSG_KEY, "prop"),
+            "89:42: " + getCheckMessage(MSG_KEY, "prop2"),
+        };
+        verifyWithInlineConfigParser(
+                 getPath("InputHiddenField3PropertySetter.java"), expected);
+    }
+
+    @Test
+    public void testHiddenField3Enum() throws Exception {
+        final String[] expected = {
+            "30:17: " + getCheckMessage(MSG_KEY, "hidden"),
+            "40:31: " + getCheckMessage(MSG_KEY, "hidden"),
+            "47:13: " + getCheckMessage(MSG_KEY, "hidden"),
+            "53:13: " + getCheckMessage(MSG_KEY, "hiddenStatic"),
+        };
+        verifyWithInlineConfigParser(
+                 getPath("InputHiddenField3Enum.java"), expected);
+    }
+
+    @Test
+    public void testHiddenField3Misc() throws Exception {
+        final String[] expected = {
+            "18:45: " + getCheckMessage(MSG_KEY, "x"),
+            "23:34: " + getCheckMessage(MSG_KEY, "xAxis"),
+        };
+        verifyWithInlineConfigParser(
+                 getPath("InputHiddenField3Misc.java"), expected);
     }
 
     @Test
